Move mapping to data.go
