+++ metricbeat/module/redis/slowlog/slowlog.go
@@ -33,16 +33,6 @@ var (
 	debugf = logp.MakeDebug("redis-slowlog")
 )
 
-// log contains all data related to one slowlog entry
-type log struct {
-	id        int64    // A unique progressive identifier for every slow log entry.
-	timestamp int64    // The unix timestamp at which the logged command was processed.
-	duration  int      // The amount of time needed for its execution, in microseconds.
-	cmd       string   // The array composing the arguments of the command.
-	key       string   // Client IP address and port (4.0 only).
-	args      []string // Client name if set via the CLIENT SETNAME command (4.0 only).
-}
-
 func init() {
 	mb.Registry.MustAddMetricSet("redis", "slowlog", New,
 		mb.WithHostParser(parse.PassThruHostParser),
@@ -81,61 +71,15 @@ func New(base mb.BaseMetricSet) (mb.MetricSet, error) {
 	}, nil
 }
 
-// Fetch fetches metrics from Redis by issuing the SLOWLOG command.
+// Fetch fetches slow operations from Redis by issuing the SLOWLOG GET command.
 func (m *MetricSet) Fetch() ([]common.MapStr, error) {
-	// Issue slowlog len.
-	count, err := redis.FetchSlowLogCount(m.pool.Get())
-	if err != nil {
-		return nil, err
-	}
-
 	// Issue slowlog get
 	slowlogs, err := redis.FetchSlowLog(m.pool.Get())
 	if err != nil {
 		return nil, err
 	}
 
-	var events []common.MapStr
-	for _, item := range slowlogs {
-		entry, err := rd.Values(item, nil)
-		if err != nil {
-			logp.Err("Error loading slowlog values: %s", err)
-			continue
-		}
-
-		var log log
-		var args []string
-		rd.Scan(entry, &log.id, &log.timestamp, &log.duration, &args)
-
-		// This splits up the args into cmd, key, args.
-		argsLen := len(args)
-		if argsLen > 0 {
-			log.cmd = args[0]
-		}
-		if argsLen > 1 {
-			log.key = args[1]
-		}
-
-		// This could contain confidential data, processors should be used to drop it if needed
-		if argsLen > 2 {
-			log.args = args[2:]
-		}
-
-		event := common.MapStr{
-			"id":  log.id,
-			"cmd": log.cmd,
-			"key": log.key,
-			"duration": common.MapStr{
-				"us": log.duration,
-			},
-		}
-
-		if log.args != nil {
-			event["args"] = log.args
-		}
-
-		events = append(events, event)
-	}
+	events := eventMapping(slowlogs)
 
 	debugf("Redis SLOWLOG from %s: %+v", m.Host(), events)
 	return events, nil
